move integration under strands folder to allow for other providers in the future
