+++ src/bedrock_agentcore/memory/integrations/strands/README.md
@@ -36,8 +36,8 @@ import boto3
 from datetime import date
 from strands import Agent
 from bedrock_agentcore.memory import MemoryClient
-from bedrock_agentcore.memory.integrations.config import AgentCoreMemoryConfig, RetrievalConfig
-from bedrock_agentcore.memory.integrations.session_manager import AgentCoreMemorySessionManager
+from bedrock_agentcore.memory.integrations.strands.strands.strands.strands.strands.config import AgentCoreMemoryConfig, RetrievalConfig
+from bedrock_agentcore.memory.integrations.strands.session_manager import AgentCoreMemorySessionManager
 ```
 
 ### Create a Basic Memory
@@ -99,8 +99,8 @@ agent("What should I buy for lunch today?")
 ### Create LTM Memory with Strategies
 
 ```python
-from bedrock_agentcore.memory.integrations.config import AgentCoreMemoryConfig, RetrievalConfig
-from bedrock_agentcore.memory.integrations.session_manager import AgentCoreMemorySessionManager
+from bedrock_agentcore.memory.integrations.strands.config import AgentCoreMemoryConfig, RetrievalConfig
+from bedrock_agentcore.memory.integrations.strands.session_manager import AgentCoreMemorySessionManager
 from datetime import datetime
 
 # Create comprehensive memory with all built-in strategies
